requirements.txt
apispec==6.0.2
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Bcrypt==1.0.1
flask-smorest==0.40.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
psycopg2-binary==2.9.5
python-decouple==3.7
SQLAlchemy==2.0.2
typing_extensions==4.4.0
webargs==8.2.0
Werkzeug==2.2.2
zipp==3.11.0
